refactor(server): extract ServiceContainer registration to boot/boot-services.ts (#4243 step 6)
Extract registerCoreServices() from server.ts main() into
boot/boot-services.ts. Registers sessionManager, authManager,
channelManager, and sessionMonitor with lifecycle hooks.

Keeps the auth token auto-repair logic co-located with authManager
registration (#3356/#3484/#3567).

Changes:
- New: src/boot/boot-services.ts (104 lines)
- Modified: src/server.ts (889 → 826, -63 lines)
- Removed unused imports: getAuthTokenFilePath, persistAuthTokenFile,
  existsSync, readFileSync
- 5683 tests passed

Part of #4243.
